+++ CHANGELOG.md
@@ -1,7 +1,15 @@
 # FastAPI Starter - Clean Architecture
 Proyecto Starter para Python FastAPI utilizando clean architecture
 
-### [V0.0.2] Crear Estructura - 2026-2-20
+### [V0.0.3] Guardar datos de Formulario - 2026-2-20
+- **Pruebas:** ✅ Todo ok.
+- **Pruebas:** ✅ Implementar formulario y recibir esos datos introducidos en una lista Python.
+- **Pruebas:** ✅ Probar introducción de datos y recepción de datos en la lista.
+- **HTML:** Añadir código para formulario  `/templates/index.html`
+- **CSS:** Añadir estilos 
+
+
+### [V0.0.2] Añadir Jinja2 - 2026-2-20
 - **Pruebas:** ✅ Todo ok.
 - **Pruebas:** ✅ Utilizar `Jinja2`.
 - **Pruebas:** ✅ Probar el acceso y renderizado de datos dentro de un diccionario.